requirements.txt
streamlit
tensorflow
pillow
numpy
requests
protobuf==3.20.*